Add new walk style
The male version of femme is not included as a walk.ing style, this will add it.

As a note, I went trough the list and tried all the walks to make sure that this walk wasn't already included under another name. I could not find any that were the same, if there is one then please lmk.